blockchain: Move LoadUtxoDB to UtxoBackend.
This moves LoadUtxoDB and the associated helper functions and variables
from utxodb.go to utxobackend.go since LoadUtxoDB is specific to the
LevelDbUtxoBackend implementation of the UtxoBackend interface.
